move watch server to own file
